fix(store-service): refactor , testcode
- api 리팩토링
- testcode 작성
