+++ store-service/src/docs/asciidoc/api-docs.adoc
@@ -70,12 +70,27 @@ operation::item-get[snippets='curl-request,http-request,http-response,path-param
 === 상품 조회 (존재하지 않는 상품)
 operation::item-get-notExistItemException[snippets='curl-request,http-request,http-response,path-parameters,response-fields']
 
+
+=== 상품 조회(판매자)
+operation::owner-item-get[snippets='curl-request,http-request,http-response,path-parameters,response-fields']
+=== 상품 리스트 조회
+operation::owner-itemList-get[snippets='curl-request,http-request,http-response,response-fields']
+
+=== 상품 수정
+operation::owner-put-item[snippets='curl-request,http-request,http-response']
+=== 상품 등록
+operation::owner-create-item[snippets='curl-request,http-request,http-response']
+
 == 카테고리
 == 카테고리 조회
 operation::get-categoryList[snippets='curl-request,http-request,http-response,request-parameters,response-fields']
 
 == 카테고리 수정
 operation::put-categoryList[snippets='curl-request,http-request,http-response,request-body,request-fields']
+
+
+
+
 
 == 매장
 === 매장 검색 조회
